requirements.txt
langchain
langchain-community
faiss-cpu
sentence-transformers
pymupdf
python-dotenv
streamlit
